Implement the module install function
This function can inject a module to a specific elf file by path.

Since opening the executing with O_RDWR will get the `Text file busy`
error, this implements it by unlinking, recreating and patching the
file.

The kernel module is inserted into the redundant space in the back
of the file and is aligned to PAGE_SIZE, which is needed by mmap.

This function will patch the entry point, and adjust some offset
in the shellcode, ensuring it jumps to the original entry point
after inserting the module.