Add get syscall function by number utilities
This change adds the function `get_syscall` to get the address of
syscall function with syscall number. This function can work without
kprobe and kallsym support.

It first finds the address of x64_sys_call and hooks all the functions
potentially being called in it by disassembling limited assembly codes.
Then, in the evil hook function, it resolves the address where it is
called from and finds out the address of syscall functions.