Find x64_sys_call by counting call instructions
Previously, the `x64_sys_call` is resolved by going up a fixed length
path on the call trace. However, with different kernel versions or
compilers, the length may differ. This counts the `call` instructions
used in the function to distinguish the `x64_sys_call` since it calls
lots of syscall functions.